feat(rust): add export_encrypted_maps_canister! macro and dogfood it
Part of #402. The reference EncryptedMaps canister was ~200 lines of
mechanical boilerplate (msg_caller injection, ByteBuf<->Blob conversion,
thread_local state, one wrapper per library method) that every adopter
copies. This adds a library macro that generates the whole canister:

    ic_vetkeys::export_encrypted_maps_canister!("my_domain_separator");
    ic_cdk::export_candid!();

Because the macro is the single source of the endpoint set, the exposed
Candid interface is exactly the one the @icp-sdk/vetkeys frontend expects,
by construction.

- New macro in ic_vetkeys::encrypted_maps (exported at crate root).
- Reference canister reduced to the macro invocation (dogfooding).

Verified locally:
- The macro-generated .did is byte-identical to the previous hand-written
  canister's .did.
- All 18 existing canister integration tests pass against the
  macro-generated wasm.
- rustfmt + clippy clean.

export_candid!() stays at the call site because ic_cdk::export_candid!
cannot be expanded from within another macro.
